+++ src/components/profile-form/profile-form.tsx
@@ -1,12 +1,12 @@
 import styles from './profile-form.module.css'
 import EditInput from '../input/edit-input/edit-input'
-import useInput from '../../services/customHooks/useInput'
+import useInput from '../../services/custom-hooks/use-input'
 import { useDispatch } from 'react-redux'
 import { Button } from '@ya.praktikum/react-developer-burger-ui-components'
 import { updateUser } from '../../services/actions/thunk/user'
 import { useCallback, useState } from 'react'
 import type { TOnSubmitCallback } from '../../utils/types'
-import useAppSelector from '../../services/customHooks/useAppSelector'
+import useAppSelector from '../../services/custom-hooks/use-app-selector'
 
 export interface IDisableHandleData {[key: string]: boolean}
