+++ src/components/protected-auth-route.tsx
@@ -1,6 +1,6 @@
 import { Navigate, Outlet, useLocation } from 'react-router-dom'
 import { FC } from 'react'
-import useAppSelector from '../services/customHooks/useAppSelector'
+import useAppSelector from '../services/custom-hooks/use-app-selector'
 
 const ProtectedAuthRoute: FC = () => {
   const { state } = useLocation()
